requirements.txt
airtest
pocoui
airtest_selenium
pytest
pynput
allure-pytest
openpyxl
pyautogui
pytest-rerunfailures
selenium==3.14
opencv-python
jsonpath
pyinstaller